% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iBioGen.R
\name{sim_tree}
\alias{sim_tree}
\title{Simulate a phylogeny with abundance and genetic diversity at the tips}
\usage{
sim_tree(simulation_name = "my_sim", project_dir = "./default_iBioGen",
  birth_rate = 1, stop_criterion = "taxa", ntaxa = 20, time = 4,
  process = "rate", ClaDS = FALSE, abundance_mean = 5e+05,
  abundance_sigma = 0.1, growth_rate_mean = 0,
  growth_rate_sigma = 0.01, ClaDS_sigma = 0.9, ClaDS_alpha = 0.1,
  sequence_length = 1000, mutation_rate = 1e-07, sample_size = 10,
  abundance_scaling = "None", nsims = 1, quiet = FALSE)
}
\arguments{
\item{simulation_name}{The name of this simulation scenario}

\item{project_dir}{Where to save files.}

\item{birth_rate}{Speciation rate}

\item{stop_criterion}{Whether to stop on ntaxa or time}

\item{ntaxa}{Number of taxa to simulate if sto is \code{ntaxa}}

\item{time}{Amount of time to simulate if stop is \code{time}}

\item{process}{Whether to evolve \code{abundance} or growth \code{rate} via BM}

\item{ClaDS}{Whether to allow speciation rates to change along the branches a la ClaDS}

\item{abundance_mean}{Ancestral abundance at time 0}

\item{abundance_sigma}{Rate at which abundance changes if process is \code{abundance}}

\item{growth_rate_mean}{Ancestral population growth rate at time 0}

\item{growth_rate_sigma}{Rate at which growth rate changes if process is \code{rate}}

\item{ClaDS_sigma}{Rate at which speciation rate changes if ClaDS is True}

\item{ClaDS_alpha}{Rate shift if ClaDS is True}

\item{sequence_length}{Length of the genomic region simulated, in base pairs}

\item{mutation_rate}{Mutation rate per base per generation}

\item{sample_size}{Number of samples to draw for calculating genetic diversity}

\item{abundance_scaling}{Scaling abundance to Ne. Can be None, log, ln or a ratio}

\item{nsims}{Number of independent simulations to perform}

\item{quiet}{Suppress printing the progress bar}
}
\description{
Simulate a phylogeny with abundance and genetic diversity at the tips
}
\details{
A birth/death process with abundances and genetic diversity at the tips.
Abundance can evolve either as a BM process with random fission at speciation
events, or the rate of change (r) of abundance can evolve as BM in which case
abundance (n) changes through time (dt) via n * exp(r*dt). Speciation rate
can also shift at branching events in the manner of ClaDS.

The iBioGen package is a product of the iBioGen project, an EU-funded
(Horizon 2020) project, focused on island biodiversity genomics. More details
about the project can be found on the website: https://www.ibiogen.eu/
}
\examples{
sim_tree() 
sim_tree(ClaDS=TRUE, ntaxa=200)
}
